AI Products/CSR Gen AI/requirements.txt
mistralai
google-genai
python-dotenv
pydantic
google-adk
pypandoc
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
